Respect read repair specifications
Summary:
In order to avoid resurrections, we only want to purge repaired
tombstones. The read repair mechanism causes old
tombstones to be replicated once compaction removes a tombstone on
one node, and this replicated tombstone is written to an unrepaired
sstable when it is flushed. This blocks the garbage collection
of any newer tombstones, which effectively stops most tombstone
garbage collection and leads to an unbounded accumulation of
old tombstones.

Unfortunately, read repair cannot be completely
turned off in the current code because the read repair mechanism
does not always respect the configuration specified in the table
schema. This diff fixes this by plumbing read repair
decisions through to where the repair is actually made, and in some
cases newly adds read repair decisions to places in the code that were
unconditionally issuing read repairs before this change.

Test Plan:
Manual testing to verify that there are no old tombstones
substantially older than gc_grace_seconds or
tombstone_compaction_interval when the schema specifies
only_purge_repaired tombstones and switches
off read repair. Before a package containing this diff is used in
production, we will run build sdmain to test for regressions and
add tests that expose resurrection and check for accumulation of
tombstones.

Reviewers: sp, #gatekeeper, nitro

Reviewed By: sp, #gatekeeper, nitro

Subscribers: soham

Tags: #callisto

JIRA Issues: CDM-45239

Differential Revision: http://phabricator.scale-data.com/D17940