Ignore all FSReadErrors when reading hints files
Summary:
Due to unknown reasons, hints files are sometimes corrupted.
D19299 addressed only the case of digest mismatches, but it
appears some customer clusters experience other FSReadErrors
that block cassandra from starting. Since we do not rely on
hints for correctness, this blockage is causing negative value
since it effectively kills a node until support addresses the
issue and it does not provide significant additional protection
against metadata corruption. Therefore the most cautious approach
seems to be to ignore all such errors.

This is intended to mitigate this problem until we are able to
find the root cause of why the hints files are corrupt.

Test Plan:
manually corrupt a hints file to reproduce the
crash loop and verify that restarting cassandra with the
new binary detects the problem, but ignores it and proceeds
as usual, clearing the corrupt hints file.

Reviewers: nitro, #gatekeeper

Reviewed By: nitro, #gatekeeper

Subscribers: soham

Tags: #callisto

JIRA Issues: CDM-51176

Differential Revision: http://phabricator.scale-data.com/D19331