Ignore digest exceptions when reading hints files
Summary:
Backport of D19299 from rubrik-3.0.9 to 3.0.9-rubrik-2.
Due to unknown reasons, hints files are sometimes corrupted.
This causes cassandra to crash loop because as soon as cassandra
restarts, it tries to send hints. This diff catches such failures
and instead ignores all remaining hints in affected files, allowing
them to be deleted as if the hints had been sent normally.

This is intended to mitigate this problem until we are able to
find the root cause of why the hints files are corrupt.

Test Plan:
manually corrupt a hints file to reproduce the
crash loop and verify that restarting cassandra with the
new binary detects the problem, but ignores it and proceeds
as usual, clearing the corrupt hints file.

Reviewers: nitro, #gatekeeper

Reviewed By: nitro, #gatekeeper

Subscribers: soham

Tags: #callisto

JIRA Issues: CDM-51176

Differential Revision: http://phabricator.scale-data.com/D19489