Run repair even when the replication factor is 1
Summary:
With only_purge_repaired_tombstones set to true, we need to
ensure that repair runs so that the number of tombstones will
not grow without bound.

Test Plan:
Manually test that this allows repair to run and mark the resulting
sstables as repaired. Will run build sdmain to test for regressions
in the non-edge case.

Reviewers: nitro, mc, #gatekeeper, fb

Reviewed By: mc, #gatekeeper, fb

Subscribers: soham

Tags: #callisto

JIRA Issues: CDM-53892

Differential Revision: http://phabricator.scale-data.com/D19627