Ignore preallocated commitlog segments on replay
Summary:
Cassandra ignores errors on replay for the
last commitlog segment.

We suspect that the last segment could
actually be all zeroes except for the
header in some cases, with the actual
error in the last sync section of the
last but one commitlog segment. This could
happen because the CommitLogSegmentManager
appears to preallocate commitlog segments.
On our internal clusters we often see that
the last but one commitlog segment is being
written to, while the last commitlog segment
is all zeroes except for the header.

To prevent Cassandra startup from failing
in such cases, we should ignore preallocated
commitlog segments starting from the end
during replay.

Also, modify the logic to skip totally
empty commitlog segments to only skip them
if they are at the end.

Test Plan:
Kill Cassandra, generate commitlog segments
at the end either all zeroed or with
everything but the header zeroed. Verify that
recovery skips these commitlogs.

Reviewers: jd, nitro

Differential Revision: https://phabricator.rubrik.com/D26446