Allow an empty sync section in the last segment
Summary:
The CommitLogSegmentManager performs a sync on clean shutdown.
In such cases, the extra preallocated segment will have an
extra sync marker written after the header. The sync section
itself not contain any mutations. If this is the case, we
expect all bytes after the sync marker to be zero. On unclean
shutdown, we do not expect the extra sync marker and the earlier
logic still holds.

Example commit log after clean shutdown:
00000000  00 00 00 06 00 00 01 5d  21 54 68 e2 00 02 7b 7d  |.......]!Th...{}|
00000010  e2 ed 2e 5d 00 00 00 1c  f7 d3 1e 15 00 00 00 00  |...]............|

The descriptor header takes 20 bytes and the sync marker
consists of the next 8.

Test Plan:
`svc -d` cassandra, and run replay. The last commitlog
segment was being considered for replay. After the change, it
was skipped.

Reviewers: jd, grammar-police!, nitro

Reviewed By: jd, nitro

Differential Revision: https://phabricator.rubrik.com/D26642