Crash on error while writing to commitlog
Summary:
Exit if there are any exceptions while adding a
mutation to the commitlog. The effect of this
should be that at most the last entry in the
last commit log is corrupted, and that entry
can be safely ignored because it was not
acknowledged.

This is the part of
https://phabricator.rubrik.com/D25977

that we need to land. The rest of the fixes
@jd suggested in the diff are checked in
as part of separate diffs.

Test Plan: deploy on scale + gin/uptime

Reviewers: jd, nitro

Reviewed By: jd

Tags: #callisto

JIRA Issues: CDM-68323

Differential Revision: https://phabricator.rubrik.com/D27180