Tolerate replay errors only in the last sync section
Summary:
From a reading of the code, the existing logic seems
to be incorrectly tolerant of commitlog errors. It is
not valid to ignore errors except for those in the
last sync section, since mutations in earlier sections
have been sync-ed and thus probably ack-ed to a client.

Test Plan:
manual, review

Manually corrupt the last but one sync section and
check that replay fails. Before the diff, it should pass.

Reviewers: jd, nitro

Reviewed By: jd

Differential Revision: https://phabricator.rubrik.com/D26514