Relax checksum failure tolerance condition on replay
Summary:
Previously, we were tolerating mutation checksum
failures only in the last sync section of the last
commitlog segment of the _first_ commitlog run.
However, on crystal, we see multiple commitlog runs,
with the checksum failure in a commitlog run which
is not the first.

We don't have a good theory yet why multiple commitlog
runs exist, but they could be related to crashes
during commitlog replay, or possibly gossip (have not
read the code to verify this).

Bump the Cassandra version with this change.

Test Plan: manual, review

Reviewers: jd, nitro

Reviewed By: jd

Tags: #callisto

JIRA Issues: CDM-70954

Differential Revision: https://phabricator.rubrik.com/D28699