Revert "Upgrade netty version for security"
Summary:
This reverts commit 6713a757ddd725be8abcba6188d57e8ec10fbb28.

This is needed because cqlsh queries have
been observed to spuriously timeout and fail after this
change (see CDM-75755). Revert this for now rather than
understand the root cause because this change is needed
for the 4.1-DA release.

Test Plan:
Will run acceptance with new Cassandra jar after the
next diff which will bump the Cassandra version, also
will verify that the timeouts and failures go away with
this change.

Reviewers: jd, fb

Reviewed By: fb

Subscribers: dar

Tags: #callisto

JIRA Issues: CDM-75755

Differential Revision: https://phabricator.rubrik.com/D34020