Connect to cqlproxy in cqlsh based on indicator file
Summary:
On rubrik platforms, the presence of /var/lib/rubrik/cqlproxy_enabled
indicates that all cassandra queries should be routed to cqlproxy.
Implement this in rubrik's fork of cqlsh.

Test Plan:
Manually test that the presence of this file makes cqlsh connect
to 27577 and its absence makes it connect to 9042 normally.

Reviewers: arijit, #security-leads

Reviewed By: arijit, #security-leads

Differential Revision: https://phabricator.rubrik.com/D38682