Make sure we should hint before hinting
Summary:
Hints accumulate for cas queries when one node is down in the Cassandra
cluster because currently the shouldHint helper method is not being
checked in this code path. This diff adds a shouldHint check for these
paths to help ensure we do not accumulate hints.

This does not completely address the issue with accumulating hints
because when a cassandra node restarts, the downtime for other nodes in
the cluster is reset to zero, so hints will begin accumulating for
another hint-max-window if one of the nodes other than the down node
restarts. Also, there are a few other paths that can create hints via
HintService#write. We will punt on this and wait for datastax to
hopefully put in a better fix because fixing this additional aspect
of hint accumulation appears more complicated (issue has been
reported as CASSANDRA-14215).

We also include a conf so that this change can effectively be reverted
via a configuration change to cassandra-env.sh.

Test Plan:
Will run jar on sazerac and see if hints accumulate as expected
for only three hours.
Will run build sdmain and acceptance for the sdmain diff after
landing this and releasing.

Reviewers: arijit, Alfred

Reviewed By: arijit

Tags: #callisto

JIRA Issues: CDM-82518

Differential Revision: https://phabricator.rubrik.com/D40462